Swapped names of DetailsView and GetDetailsView for more consistent naming convention (DetailsView isn't used)
