+++ src/main/java/app/WeatherAppBuilder.java
@@ -22,12 +22,8 @@ import use_case.get_details.GetDetailsInputBoundary;
 import use_case.get_details.GetDetailsInterator;
 import use_case.get_details.GetDetailsOutputBoundary;
 import use_case.get_forecast.GetForecastInteractor;
-import data_access.FavoriteCityStorageImpl;
 import entity.*;
 
-import interface_adapter.manage_cities.ManageCitiesController;
-
-import use_case.manage_cities.FavoriteCitiesInteractor;
 import use_case.search_city.SearchCityInputBoundary;
 import use_case.search_city.SearchCityInteractor;
 
@@ -37,22 +33,15 @@ import interface_adapter.manage_sort.SortCitiesViewModel;
 import interface_adapter.nearby_cities.NearbyCitiesController;
 import interface_adapter.nearby_cities.NearbyCitiesPresenter;
 import interface_adapter.nearby_cities.NearbyCitiesViewModel;
-import interface_adapter.search_city.SearchCityPresenter;
-import interface_adapter.search_city.SearchCityViewModel;
 import use_case.get_forecast.GetForecastInputBoundary;
-import use_case.get_forecast.GetForecastInteractor;
 import use_case.get_forecast.GetForecastOutputBoundary;
-import entity.*;
 
 import use_case.nearby_cities.NearbyCitiesInputBoundary;
 import use_case.nearby_cities.NearbyCitiesInteractor;
 import use_case.nearby_cities.NearbyCitiesOutputBoundary;
-import use_case.search_city.SearchCityOutputBoundary;
-import view.*;
 import interface_adapter.get_forecast.GetForecastController;
 import interface_adapter.get_forecast.GetForecastPresenter;
 import interface_adapter.get_forecast.GetForecastViewModel;
-import data_access.ForecastWeatherInfoObject;
 
 public class WeatherAppBuilder {
     private final JPanel cardPanel = new JPanel();
@@ -82,7 +71,7 @@ public class WeatherAppBuilder {
     private SearchCityView searchCityView;
 
     private GetDetailsViewModel getDetailsViewModel;
-    private DetailsView getDetailsView;
+    private GetDetailsView getDetailsView;
 
     public WeatherAppBuilder() {
         cardPanel.setLayout(cardLayout);
@@ -154,7 +143,7 @@ public class WeatherAppBuilder {
      */
     public WeatherAppBuilder addGetDetailsView() {
         getDetailsViewModel = new GetDetailsViewModel();
-        getDetailsView = new DetailsView(getDetailsViewModel);
+        getDetailsView = new GetDetailsView(getDetailsViewModel);
         cardPanel.add(getDetailsView, getDetailsView.getViewName());
         return this;
     }
